Fixed passive Floater and import simple water
Passive Floater now uses the mass of the floated objects to apply the
lifting force,
if the object sinks or floats and how high it floats is now determined
by it's mass.
Still using the objects origin/Transform to apply the force so they
rotate like if they where in space.